Feature/estabelecimento crud

+++ src/main/java/com/controle_de_gastos/notas_api/service/EstabelecimentoService.java
@@ -1,13 +1,11 @@
 package com.controle_de_gastos.notas_api.service;
 
-import com.controle_de_gastos.notas_api.dto.resposta.EstabelecimentoBairroRespostaDTO;
 import com.controle_de_gastos.notas_api.repository.CategoriaEstabelecimentoRepository;
 import com.controle_de_gastos.notas_api.repository.EstabelecimentoBairroJuncaoRepository;
 import com.controle_de_gastos.notas_api.repository.EstabelecimentoRepository;
 import com.controle_de_gastos.notas_api.dto.projecao.BairroComEnderecoProjecaoDTO;
 import com.controle_de_gastos.notas_api.dto.resposta.EstabelecimentoComBairroRespostaDTO;
 import com.controle_de_gastos.notas_api.dto.resposta.EstabelecimentoRespostaDTO;
-import com.controle_de_gastos.notas_api.dto.projecao.EstabelecimentoComEnderecoProjecaoDTO;
 import com.controle_de_gastos.notas_api.dto.requisicao.EstabelecimentoRequisicaoDTO;
 import com.controle_de_gastos.notas_api.model.Estabelecimento;
 import com.controle_de_gastos.notas_api.model.CategoriaEstabelecimento;
@@ -36,15 +34,6 @@ public class EstabelecimentoService {
         );
     }
 
-    public EstabelecimentoComEnderecoProjecaoDTO toEstabSimplesDTO(Estabelecimento estabelecimento, String endereco) {
-        return new EstabelecimentoComEnderecoProjecaoDTO(
-                estabelecimento.getId(),
-                estabelecimento.getNome(),
-                endereco
-        );
-    }
-
-
     public List<EstabelecimentoRespostaDTO> listarTodos(){
         return estabelecimentoRepository.findAll()
                 .stream()
@@ -52,11 +41,11 @@ public class EstabelecimentoService {
                 .toList();
     }
 
-    public EstabelecimentoComBairroRespostaDTO listarBairroPorEstabelecimentoId(Integer id) {
+    public Optional<EstabelecimentoComBairroRespostaDTO> listarBairroPorEstabelecimentoId(Integer id) {
                List <EstabelecimentoBairroJuncao> juncoes = estabelecimentoBairroJuncaoRepository.findByEstabelecimentoId(id);
 
                if(juncoes.isEmpty()){
-                   return null;
+                   return Optional.empty();
                }
 
                Estabelecimento estabelecimento = juncoes.get(0).getEstabelecimento();
@@ -68,11 +57,11 @@ public class EstabelecimentoService {
                        ))
                        .toList();
 
-                return new EstabelecimentoComBairroRespostaDTO(
+                return Optional.of(new EstabelecimentoComBairroRespostaDTO(
                         estabelecimento.getId(),
                         estabelecimento.getNome(),
                         bairros
-                );
+                ));
     }
 
     public List<EstabelecimentoComBairroRespostaDTO> listarTodosBairroPorEstabelecimento(){
@@ -104,16 +93,16 @@ public class EstabelecimentoService {
                     .toList();
 
             return juncaos;
-
     }
 
     public EstabelecimentoRespostaDTO criar(EstabelecimentoRequisicaoDTO estabelecimentoDto) {
         CategoriaEstabelecimento categoria = categoriaEstabelecimentoRepository.findById(estabelecimentoDto.idCategoria())
                 .orElseThrow(()-> new RuntimeException("Categoria Inválida"));
 
-        Estabelecimento estabelecimento = new Estabelecimento();
-        estabelecimento.setCategoria(categoria);
-        estabelecimento.setNome(estabelecimentoDto.nome());
+        Estabelecimento estabelecimento = Estabelecimento.builder()
+                .categoria(categoria)
+                .nome(estabelecimentoDto.nome())
+                .build();
         return toRespostaDTO(estabelecimentoRepository.save(estabelecimento));
     }
 
@@ -122,8 +111,27 @@ public class EstabelecimentoService {
                 .map(this::toRespostaDTO);
     }
 
-    public void deletarPorId(Integer id){
-        estabelecimentoRepository.deleteById(id);
+    public Optional<EstabelecimentoRespostaDTO> atualizarTudo(Integer id, EstabelecimentoRequisicaoDTO estabelecimentoDto) {
+        Estabelecimento estabelecimento = estabelecimentoRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Estabelecimento não encontrado"));
+
+        CategoriaEstabelecimento catEstab = categoriaEstabelecimentoRepository.findById(estabelecimentoDto.idCategoria())
+                .orElseThrow(()-> new RuntimeException("Categoria não Encontrada"));
+
+        estabelecimento.setNome(estabelecimentoDto.nome());
+        estabelecimento.setCategoria(catEstab);
+        return Optional.of(toRespostaDTO(estabelecimentoRepository.save(estabelecimento)));
+    }
+
+    public boolean deletarPorId(Integer id){
+        Optional<Estabelecimento> estabelecimento = estabelecimentoRepository.findById(id);
+        if (estabelecimento.isEmpty()) return false;
+
+        if(!estabelecimento.get().getEstabelecimentoBairroJuncaos().isEmpty()){
+            throw new IllegalStateException("Não é possível deletar, existem dependências");
+        }
+                estabelecimentoRepository.deleteById(id);
+        return true;
     }
